pack: Handle size boundary of extending msgpack map

+++ include/fluent-bit/flb_pack.h
@@ -111,8 +111,8 @@ flb_sds_t flb_msgpack_raw_to_json_sds(const void *in_buf, size_t in_size, int es
 
 int flb_pack_time_now(msgpack_packer *pck);
 int flb_msgpack_expand_map(char *map_data, size_t map_size,
-                           msgpack_object_kv **obj_arr, int obj_arr_len,
-                           char** out_buf, int* out_size);
+                           msgpack_object_kv **obj_arr, size_t obj_arr_len,
+                           char** out_buf, size_t *out_size);
 
 struct flb_gelf_fields {
     flb_sds_t timestamp_key;
